+++ source/tokenizer.h
@@ -19,10 +19,10 @@ enum TokenKind
 // structure for token
 struct Token
 {
-    TokenKind kind; // kind of token
-    char *str;      // pointer to token string
-    size_t len;     // length of token string
-    uint32_t value; // value of token (only for TK_IMMEDIATE)
+    TokenKind kind;  // kind of token
+    char *str;       // pointer to token string
+    size_t len;      // length of token string
+    uintmax_t value; // value of token (only for TK_IMMEDIATE)
 };
 
 bool peek_reserved(const char *str);
